% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_all_data}
\alias{extract_all_data}
\title{Extract important data from Geyer et al. 2017}
\usage{
extract_all_data()
}
\arguments{
\item{location}{URL of pdf containing raw data.}
}
\value{
A list object with 4 items. Each item is an extracted data frame.
}
\description{
Extract important data from Geyer et al. 2017
}
\examples{

#Extract important data from Geter et al. 2017
all_data <- extract_all_data()
}
