% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{sectors_plastic}
\alias{sectors_plastic}
\title{Extract table with proportion of plastic resins used in each sector.}
\usage{
sectors_plastic(extracted_data)
}
\arguments{
\item{extracted_data}{Raw scraped data generated by package tabulizer.}
}
\value{
A dataframe.
}
\description{
Extract table with proportion of plastic resins used in each sector.
}
