% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{sectors_waste}
\alias{sectors_waste}
\title{Extract table with 2015 plastic production and waste per sector.}
\usage{
sectors_waste(extracted_data)
}
\arguments{
\item{extracted_data}{Raw scraped data generated by package tabulizer.}
}
\value{
A dataframe.
}
\description{
Extract table with 2015 plastic production and waste per sector.
}
