% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{standard_names}
\alias{standard_names}
\title{Names of plastic resins and sector types}
\format{A list with 2 items.
\describe{
  \item{plastics}{Standard names of plastic resin types.}
  \item{sectors}{Standard names of industry sectors.}
  }}
\description{
Standard names for each plastic resin and sector type
used in Geyer et al. 2017.

Useful data extracted from Geyer et al. 2017.
}
