% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{total_plastic}
\alias{total_plastic}
\title{Extract table with total plastic production over time}
\usage{
total_plastic(extracted_data)
}
\arguments{
\item{extracted_data}{Raw scraped data generated by package tabulizer.}
}
\value{
A dataframe.
}
\description{
Extract table with total plastic production over time
}
